create a search component for backoffice
