+++ frontend/src/containers/Backoffice.js
@@ -1,23 +1,20 @@
-import React, { useEffect, useRef, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import styled from 'styled-components'
 import BaseMap from './BaseMap'
 import LawType from '../components/backoffice/LawType'
+import SearchComponent from '../components/backoffice/SearchComponent'
 import RegulatoryZoneMetadata from '../components/regulatory_zones/RegulatoryZoneMetadata'
 import getAllRegulatoryZonesByRegTerritory from '../domain/use_cases/getAllRegulatoryZonesByRegTerritory'
 import getAllGearCodes from '../domain/use_cases/getAllGearCodes'
 import { setError } from '../domain/reducers/Global'
-import { ReactComponent as SearchIconSVG } from '../components/icons/Loupe.svg'
 import { COLORS } from '../constants/constants'
 import { BlackButton, WhiteButton } from '../components/commonStyles/Buttons.style'
 import { EmptyResult } from '../components/commonStyles/Text.style'
 import closeRegulatoryZoneMetadata from '../domain/use_cases/closeRegulatoryZoneMetadata'
 import { RegulatoryTerritory } from '../domain/entities/regulatory'
-import { search } from '../utils'
 
 const Backoffice = () => {
-  const searchInput = useRef(null)
-  const [searchText, setSearchText] = useState('')
   const [foundRegulatoryZonesByRegTerritory, setFoundRegulatoryZonesByRegTerritory] = useState({})
   const [regulatoryZoneListByRegTerritory, setRegulatoryZoneListByRegTerritory] = useState({})
   const showedLayers = useSelector(state => state.layer.showedLayers)
@@ -31,34 +28,6 @@ const Backoffice = () => {
     regulatoryZoneMetadata
   } = useSelector(state => state.regulatory)
 
-  const properties = ['layerName', 'zone', 'region', 'seafront', 'regulatoryReferences']
-
-  const searchRegulatoryZone = (searchText) => {
-    const searchResult = {}
-    if (searchText === '') {
-      setFoundRegulatoryZonesByRegTerritory(regulatoryZoneListByRegTerritory)
-    } else {
-      // by Territory
-      Object.keys(regulatoryZoneListByRegTerritory).forEach(territory => {
-        // by lawType
-        const searchResultByLawType = {}
-        Object.keys(regulatoryZoneListByRegTerritory[territory]).forEach(lawType => {
-          // make a copy
-          const regulatoryZone = Object.assign({}, regulatoryZoneListByRegTerritory[territory][lawType])
-          const foundRegulatoryZones = search(searchText, properties, regulatoryZone)
-          // Si c'est vide c'est quand même ajouté ?
-          if (foundRegulatoryZones && foundRegulatoryZones !== {}) {
-            searchResultByLawType[lawType] = foundRegulatoryZones
-          }
-        })
-        if (searchResultByLawType !== {}) {
-          searchResult[territory] = searchResultByLawType
-        }
-      })
-      setFoundRegulatoryZonesByRegTerritory(searchResult)
-    }
-  }
-
   const getRegulatoryZones = () => {
     dispatch(getAllRegulatoryZonesByRegTerritory(dispatch))
       .then(regulatoryZones => {
@@ -70,21 +39,9 @@ const Backoffice = () => {
   }
 
   useEffect(() => {
-    if (searchInput) {
-      searchInput.current.focus()
-    }
     getRegulatoryZones()
     dispatch(getAllGearCodes())
   }, [])
-
-  useEffect(() => {
-    console.log('and the result is')
-    console.log(foundRegulatoryZonesByRegTerritory)
-  }, [foundRegulatoryZonesByRegTerritory])
-
-  useEffect(() => {
-    searchRegulatoryZone(searchText)
-  }, [searchText, setFoundRegulatoryZonesByRegTerritory, regulatoryZoneListByRegTerritory])
 
   function addNewRegZone () {
     console.log('addNewRegZone clicked')
@@ -112,7 +69,7 @@ const Backoffice = () => {
   }
 
   const displayRegulatoryZoneByRegTerritory = (territory) => {
-    const territoryRegList = regulatoryZoneListByRegTerritory[territory]
+    const territoryRegList = foundRegulatoryZonesByRegTerritory[territory]
     return territoryRegList
       ? <RegulatoryZoneListByLawTypeList>{displayRegulatoryZoneListByLawType(territoryRegList)}</RegulatoryZoneListByLawTypeList>
       : <EmptyResult>Aucune zone pour ce territoire</EmptyResult>
@@ -138,19 +95,10 @@ const Backoffice = () => {
         <RegulatoryZonePanel
           regulatoryZoneMetadataPanelIsOpen={regulatoryZoneMetadataPanelIsOpen}
         >
-          <SearchContainer>
-            <SearchBoxInput
-              ref={searchInput}
-              type="text"
-              value={searchText}
-              placeholder={'Rechercher une zone par son nom ou sa référence réglementaire'}
-              onChange={e => setSearchText(e.target.value)}/>
-            <SearchButton
-              onClick={() => searchRegulatoryZone()}
-            >
-              <SearchIcon/>
-            </SearchButton>
-          </SearchContainer>
+          <SearchComponent
+            setFoundRegulatoryZonesByRegTerritory={setFoundRegulatoryZonesByRegTerritory}
+            regulatoryZoneListByRegTerritory={regulatoryZoneListByRegTerritory}
+          />
           <ButtonList>
             <WhiteButton>Brouillon (X)</WhiteButton>
             <WhiteButton>Tracé en attente (X)</WhiteButton>
@@ -226,14 +174,6 @@ const RegulatoryZoneListByLawTypeList = styled.div`
   margin: 10px 0;
 `
 
-const SearchContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  justify-content: center;
-  padding: 25px 40px 0;
-`
-
 const ButtonList = styled.div`
   display: flex;
   flex-direction:row;
@@ -279,31 +219,6 @@ const MetadataWrapper = styled.div`
   max-height: 95vh;
   transition: all 0.5s;
   opacity: ${props => props.regulatoryZoneMetadataPanelIsOpen ? '1' : '0'};
-`
-
-const SearchBoxInput = styled.input`
-  margin: 0;
-  background-color: white;
-  border: 1px ${COLORS.gray} solid;
-  border-radius: 0;
-  color: ${COLORS.grayDarkerThree};
-  font-size: 0.8em;
-  height: 40px;
-  width: 100%;
-  padding: 0 5px 0 10px;
-`
-
-const SearchIcon = styled(SearchIconSVG)`
-  width: 40px;
-  height: 40px;
-  float: right;
-  background: ${COLORS.grayDarkerThree};
-  cursor: pointer;
-`
-
-const SearchButton = styled.a`
-  width: 40px;
-  height: 40px;
 `
 
 export default Backoffice
